refactor - modify interface file name
