+++ study_rails_08/rails-react-client/src/components/main_rest_container.tsx
@@ -2,7 +2,7 @@ import React, { useState, useEffect } from 'react'
 import { UserContext } from '../global/contexts'
 
 import axios from 'axios'
-import { TweetInterface } from '../model/tweet_interface'
+import { TweetInterface } from '../model/types'
 import { TweetsContainer } from './tweets/tweets_container'
 
 export const MainRestContainer: React.FC = () => {
