+++ study_rails_08/rails-react-client/src/model/types.ts
@@ -1,4 +1,4 @@
-export interface TweetInterface {
+export type TweetInterface = {
   id: number
   sentence: string
   user_name: string
